requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
boto3==1.28.33
botocore==1.31.33
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
daphne==4.0.0
Django==4.2.3
django-admin==2.0.2
django-cors-headers==4.1.0
django-excel-response2==3.0.5
django-filter==23.2
django-jazzmin==2.6.0
django-six==1.0.5
djangorestframework==3.14.0
drf-yasg==1.21.6
excel-base==1.0.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
isoweek==1.3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.6.0
reportlab==4.0.4
requests==2.31.0
screen==1.0.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.25.4
xlwt==1.3.0
zope.interface==6.0
django-storages==1.13.2
django-json-widget==1.1.1